requirements.txt
mysql-connector-python==9.0.0
psycopg==3.2.1
psycopg-binary==3.2.1
PyYAML==6.0.2
typing_extensions==4.12.2
